removed folder grouping, moved everything from interactive into modals folder
